Add possibility to have plugins capture the next message from the user
It will now be possible to run a plugin and mark that the next chatmsg from
the same user will automatically be passed to the same plugin no matter what
was written. In this way plugins can run in a dialog with the user.

Simplified the interface for handling of messages. Result of regex is
set on the chatmsg and can be retrieved inside the handle method.

Updated composer.lock